smalll update
git-svn-id: svn+ssh://c4science.ch/source/amie/trunk@1431 64fb8e13-074e-4207-b257-2d34fb7690c6

+++ matrixops.cpp
@@ -505,7 +505,7 @@ Vector Mu::solveSystem(const Matrix & A, const Vector & b, Vector & x)
 	r = s- t*omega ;
 	double rho_ =rho ;
 	
-	int nit = 0 ;
+	size_t nit = 0 ;
 	
 	while(nit < r.size())
 	{
